requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
httpx==0.16.1
pytest==6.2.2
requests==2.25.1
pact-python==1.3.0